refactor uninstall into its own file
